server/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
gspread==6.1.0
google-auth==2.29.0
python-dotenv==1.0.1
pydantic==2.6.4
